backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pymongo==4.8.0
redis==5.0.7
boto3==1.34.162
prometheus-client==0.20.0
httpx==0.27.2
loguru==0.7.2
pytest==8.3.2
pytest-asyncio==0.23.8
